+++ History.md
@@ -5,3 +5,4 @@
 * Provides Sinatra helpers providing the same features as node version.
 * Provides Sinatra examples.
 * Provides a #reset option to empty includer's cache.
+* Defaults to bundled mode if RACK_ENV variable is set to production.